* renamed incApplSpecific/FwConvCpp32 -> incApplSpecific/applCpp32_emC Details for exception handling: * Exception: on second throw does not store the line number in throw, it is faulty. Keep the old one from the first throw. * ThreadContext: renamining to IxStacktrace_emC, stactrace->_ixStacktrace_, semanticly more exact. See in Videoclip. * getUserBuffer_ThreadContext_emC: with size aligned 8, stores 0 on freeing.

+++ source/InspcJ2c/ClassContent_Inspc.c
@@ -1127,7 +1127,7 @@ int32 cmdGetAddressByPath_ClassContent_Inspc(ClassContent_Inspc_s* thiz, struct 
       {
 
 
-        int32  addr = getMemoryIdent_FieldJc(theField, theObject, "I", idxP[0]);
+        int32  addr = (int32)getMemoryIdent_FieldJc(theField, theObject, "I", idxP[0]);
         addChildInteger_ByteDataAccessBaseJc(&((thiz->answerItem).base.super), 1, kReferenceAddr_InspcDataExchangeAccess_Inspc, _thCxt);/*Set the number of char-bytes in 1 byte*/
 
         addChildInteger_ByteDataAccessBaseJc(&((thiz->answerItem).base.super), 4, addr, _thCxt);
